connector/requirements.txt
fastapi==0.95.2
uvicorn==0.21.1
websockets==11.0.3
PyYAML==6.0
structlog==24.1.0
httpx==0.24.1
pythonnet==3.0.3